fix bags

+++ convex/files.ts
@@ -126,6 +126,13 @@ export const deleteFile = mutation({
       throw new ConvexError("You do not have access to this file")
     }
 
+    const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)
+    ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("You do not have admin access to delete this file")
+    }
+
     await ctx.db.delete(args.fileId);
   }
 });
